Use `PartitionableLoopInterface` at Flow level and on CPU backend.
This commit uses the `PartitionableLoopInterface` at the Flow level to
determine the dimensions of the op to distribute along.

The same interface is used on the CPU backend. To make this interface
easier to use some refactoring of the existing heuristics for
selecting tile sizes are refactored (mostly NFC) to reason about the
partitioned dimensions explicitly (instead of implicitly in the
code). Some utility helper methods are added to convert the
`workloadPerWorkgroup` determined into the tile sizes of the flow
level.